refactor tag actions into separate files
