+++ troposphere/static/js/actions/tag/create_AddToInstance.js
@@ -8,29 +8,6 @@ define(function (require) {
 
   return {
 
-    create: function(params){
-
-      if(!params.name) throw new Error("Missing name");
-      if(!params.description) throw new Error("Missing description");
-
-      var name = params.name,
-          description = params.description;
-
-      var tag = new Tag({
-        name: name,
-        description: description
-      });
-
-      // Add the tag optimistically
-      Utils.dispatch(TagConstants.ADD_TAG, {tag: tag}, {silent: false});
-
-      tag.save().done(function () {
-        Utils.dispatch(TagConstants.UPDATE_TAG, {tag: tag}, {silent: false});
-      }).fail(function () {
-        Utils.dispatch(TagConstants.REMOVE_TAG, {tag: tag}, {silent: false});
-      });
-    },
-
     create_AddToInstance: function(params){
 
       if(!params.name) throw new Error("Missing name");
