+++ troposphere/static/js/modals/tag/create.js
@@ -1,10 +1,7 @@
 define(function (require) {
   'use strict';
 
-  var AppDispatcher = require('dispatchers/AppDispatcher'),
-      TagConstants = require('constants/TagConstants'),
-      Tag = require('models/Tag'),
-      actions = require('actions'),
+  var actions = require('actions'),
       ModalHelpers = require('components/modals/ModalHelpers'),
       TagCreateModal = require('components/modals/tag/TagCreateModal.react');
 
@@ -18,7 +15,6 @@ define(function (require) {
 
       ModalHelpers.renderModal(modal, function(name, description){
         actions.TagActions.create({
-          initialTagName: initialTagName,
           name: name,
           description: description
         })
